Retreat to IOptionsSnapshot
