--- src/Microsoft.Extensions.Options/IOptionsManager.cs
@@ -1,17 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-namespace Microsoft.Extensions.Options
-{
-    /// <summary>
-    /// Used to retreive configured TOptions instances.
-    /// </summary>
-    /// <typeparam name="TOptions">The type of options being requested.</typeparam>
-    public interface IOptionsManager<TOptions> : IOptions<TOptions> where TOptions : class, new()
-    {
-        /// <summary>
-        /// Returns a configured TOptions instance with the given name.
-        /// </summary>
-        TOptions Get(string name);
-    }
-}
