+++ src/Microsoft.Extensions.Options/OptionsManager.cs
@@ -1,45 +1,36 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
+using System.Collections.Generic;
 
 namespace Microsoft.Extensions.Options
 {
     /// <summary>
-    /// Implementation of IOptionsFactory.
+    /// Implementation of IOptions.
     /// </summary>
-    /// <typeparam name="TOptions">The type of options being requested.</typeparam>
-    public class OptionsManager<TOptions> : IOptionsManager<TOptions>, IOptions<TOptions>, IOptionsSnapshot<TOptions> where TOptions : class, new()
+    /// <typeparam name="TOptions"></typeparam>
+    public class OptionsManager<TOptions> : IOptions<TOptions> where TOptions : class, new()
     {
-        private readonly IOptionsCache<TOptions> _cache;
-        private readonly IOptionsFactory<TOptions> _factory;
+        private LegacyOptionsCache<TOptions> _optionsCache;
 
         /// <summary>
         /// Initializes a new instance with the specified options configurations.
         /// </summary>
-        /// <param name="cache">The cache to use.</param>
-        /// <param name="factory">The factory to use to create options.</param>
-        public OptionsManager(IOptionsCache<TOptions> cache, IOptionsFactory<TOptions> factory)
+        /// <param name="setups">The configuration actions to run.</param>
+        public OptionsManager(IEnumerable<IConfigureOptions<TOptions>> setups)
         {
-            _cache = cache;
-            _factory = factory;
+            _optionsCache = new LegacyOptionsCache<TOptions>(setups);
         }
 
-        public TOptions Value
+        /// <summary>
+        /// The configured options instance.
+        /// </summary>
+        public virtual TOptions Value
         {
             get
             {
-                return Get(Options.DefaultName);
+                return _optionsCache.Value;
             }
-        }
-
-        public virtual TOptions Get(string name)
-        {
-            if (name == null)
-            {
-                throw new ArgumentNullException(nameof(name));
-            }
-            return _cache.GetOrAdd(name, () => _factory.Create(name));
         }
     }
 }
