+++ test/Microsoft.Extensions.Options.Test/OptionsSnapshotTest.cs
@@ -112,29 +112,88 @@ namespace Microsoft.Extensions.Options.Tests
         }
 
 
-        [Fact(Skip = "Snapshot is broken")]
-        public void SnapshotOptionsAreRecreatedPerScope()
+        [Fact]
+        public void SnapshotOptionsAreCachedPerScope()
         {
             var services = new ServiceCollection()
                 .AddOptions()
                 .AddScoped<IConfigureOptions<FakeOptions>, TestConfigure>()
                 .BuildServiceProvider();
 
+            var cache = services.GetRequiredService<IOptionsCache<FakeOptions>>();
             var factory = services.GetRequiredService<IServiceScopeFactory>();
             FakeOptions options = null;
+            FakeOptions namedOne = null;
             using (var scope = factory.CreateScope())
             {
                 options = scope.ServiceProvider.GetRequiredService<IOptionsSnapshot<FakeOptions>>().Value;
                 Assert.Equal(options, scope.ServiceProvider.GetRequiredService<IOptionsSnapshot<FakeOptions>>().Value);
+                namedOne = scope.ServiceProvider.GetRequiredService<IOptionsSnapshot<FakeOptions>>().Get("1");
+                Assert.Equal(namedOne, scope.ServiceProvider.GetRequiredService<IOptionsSnapshot<FakeOptions>>().Get("1"));
             }
             Assert.Equal(1, TestConfigure.ConfigureCount);
+            Assert.True(cache.TryRemove(Options.DefaultName));
+            Assert.True(cache.TryRemove("1"));
             using (var scope = factory.CreateScope())
             {
                 var options2 = scope.ServiceProvider.GetRequiredService<IOptionsSnapshot<FakeOptions>>().Value;
-                Assert.Equal(options2, scope.ServiceProvider.GetRequiredService<IOptionsSnapshot<FakeOptions>>().Value);
                 Assert.NotEqual(options, options2);
+                var namedOne2 = scope.ServiceProvider.GetRequiredService<IOptionsSnapshot<FakeOptions>>().Get("1");
+                Assert.NotEqual(namedOne2, namedOne);
             }
             Assert.Equal(2, TestConfigure.ConfigureCount);
         }
+
+        [Fact]
+        public void CanResolveNamedOptionsSnapshot()
+        {
+            var services = new ServiceCollection().AddOptions();
+
+            services.Configure<FakeOptions>("1", options =>
+            {
+                options.Message = "one";
+            });
+            services.Configure<FakeOptions>("2", options =>
+            {
+                options.Message = "two";
+            });
+
+            var sp = services.BuildServiceProvider();
+            var option = sp.GetRequiredService<IOptionsSnapshot<FakeOptions>>();
+            Assert.Equal("one", option.Get("1").Message);
+            Assert.Equal("two", option.Get("2").Message);
+        }
+
+        [Fact]
+        public void FactoryCanConfigureAllOptions()
+        {
+            var services = new ServiceCollection().AddOptions();
+            services.ConfigureAll<FakeOptions>(o => o.Message = "Default");
+
+            var sp = services.BuildServiceProvider();
+            var option = sp.GetRequiredService<IOptionsSnapshot<FakeOptions>>();
+            Assert.Equal("Default", option.Get("1").Message);
+            Assert.Equal("Default", option.Get("2").Message);
+        }
+
+        [Fact]
+        public void NamedSnapshotsConfiguresInRegistrationOrder()
+        {
+            var services = new ServiceCollection().AddOptions();
+            services.Configure<FakeOptions>("-", o => o.Message += "-");
+            services.ConfigureAll<FakeOptions>(o => o.Message += "A");
+            services.Configure<FakeOptions>("+", o => o.Message += "+");
+            services.ConfigureAll<FakeOptions>(o => o.Message += "B");
+            services.ConfigureAll<FakeOptions>(o => o.Message += "C");
+            services.Configure<FakeOptions>("+", o => o.Message += "+");
+            services.Configure<FakeOptions>("-", o => o.Message += "-");
+
+            var sp = services.BuildServiceProvider();
+            var option = sp.GetRequiredService<IOptionsSnapshot<FakeOptions>>();
+            Assert.Equal("ABC", option.Get("1").Message);
+            Assert.Equal("A+BC+", option.Get("+").Message);
+            Assert.Equal("-ABC-", option.Get("-").Message);
+        }
+
     }
 }
